[app][hwcrypto] Switch to use different kdfv1 root key
The previous implementation used RPMB authentication key
as a root for kdfv1 key derivation function. Such key is
shared with external hardware and it is unclear how well
this secret is kept by it. It is better to use secret that
is tied to device itself which is introduced by this CL.

Change-Id: I61b408da6880fac0d3013c5beac8e0ed9d53a0ab